Tidy up changes

+++ core/src/game.c
@@ -35,7 +35,7 @@ game_variant_t get_game_variant_type_from_name(const char *variant_name) {
 cgp_parse_status_t place_letters_on_board(Game *game, const char *letters,
                                           int row_start,
                                           int *current_column_index) {
-  int letters_length = string_length(letters);
+  size_t letters_length = string_length(letters);
   uint8_t *machine_letters = malloc_or_die(sizeof(uint8_t) * letters_length);
   int number_of_machine_letters =
       str_to_machine_letters(game->gen->letter_distribution, letters, false,
